+++ service/src/main/java/bio/terra/pipelines/stairway/imputation/RunImputationGcpJobFlight.java
@@ -5,9 +5,10 @@ import bio.terra.pipelines.common.utils.FlightBeanBag;
 import bio.terra.pipelines.common.utils.FlightUtils;
 import bio.terra.pipelines.common.utils.PipelinesEnum;
 import bio.terra.pipelines.dependencies.stairway.JobMapKeys;
-import bio.terra.pipelines.stairway.imputation.gcp.AddDataTableRowStep;
-import bio.terra.pipelines.stairway.imputation.gcp.PollCromwellSubmissionStatusStep;
-import bio.terra.pipelines.stairway.imputation.gcp.SubmitCromwellRunSetStep;
+import bio.terra.pipelines.stairway.imputation.steps.PrepareImputationInputsStep;
+import bio.terra.pipelines.stairway.imputation.steps.gcp.AddDataTableRowStep;
+import bio.terra.pipelines.stairway.imputation.steps.gcp.PollCromwellSubmissionStatusStep;
+import bio.terra.pipelines.stairway.imputation.steps.gcp.SubmitCromwellSubmissionStep;
 import bio.terra.stairway.Flight;
 import bio.terra.stairway.FlightMap;
 import bio.terra.stairway.RetryRule;
@@ -58,8 +59,10 @@ public class RunImputationGcpJobFlight extends Flight {
         new AddDataTableRowStep(flightBeanBag.getRawlsService(), flightBeanBag.getSamService()));
 
     addStep(
-        new SubmitCromwellRunSetStep(
-            flightBeanBag.getRawlsService(), flightBeanBag.getSamService()),
+        new SubmitCromwellSubmissionStep(
+            flightBeanBag.getRawlsService(),
+            flightBeanBag.getSamService(),
+            flightBeanBag.getImputationConfiguration()),
         dbRetryRule);
 
     addStep(
