+++ service/src/test/java/bio/terra/pipelines/stairway/imputation/steps/azure/CheckLeonardoHealthStepTest.java
@@ -1,4 +1,4 @@
-package bio.terra.pipelines.stairway.imputation;
+package bio.terra.pipelines.stairway.imputation.steps.azure;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
