+++ service/src/test/java/bio/terra/pipelines/stairway/imputation/steps/azure/CheckWdsHealthStepTest.java
@@ -1,12 +1,13 @@
-package bio.terra.pipelines.stairway.imputation;
+package bio.terra.pipelines.stairway.imputation.steps.azure;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
 
-import bio.terra.pipelines.dependencies.cbas.CbasService;
 import bio.terra.pipelines.dependencies.common.HealthCheckWorkspaceApps;
 import bio.terra.pipelines.dependencies.sam.SamService;
+import bio.terra.pipelines.dependencies.wds.WdsService;
+import bio.terra.pipelines.stairway.imputation.RunImputationJobFlightMapKeys;
 import bio.terra.pipelines.testutils.BaseEmbeddedDbTest;
 import bio.terra.pipelines.testutils.StairwayTestUtils;
 import bio.terra.stairway.FlightContext;
@@ -17,8 +18,8 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mock;
 
-class CheckCbasHealthStepTest extends BaseEmbeddedDbTest {
-  @Mock private CbasService cbasService;
+class CheckWdsHealthStepTest extends BaseEmbeddedDbTest {
+  @Mock private WdsService wdsService;
   @Mock private SamService samService;
   @Mock private FlightContext flightContext;
 
@@ -26,7 +27,7 @@ class CheckCbasHealthStepTest extends BaseEmbeddedDbTest {
   void setup() {
     FlightMap inputParameters = new FlightMap();
     FlightMap workingMap = new FlightMap();
-    workingMap.put(RunImputationJobFlightMapKeys.CBAS_URI, "cbasUri");
+    workingMap.put(RunImputationJobFlightMapKeys.WDS_URI, "wdsUri");
 
     when(flightContext.getInputParameters()).thenReturn(inputParameters);
     when(flightContext.getWorkingMap()).thenReturn(workingMap);
@@ -35,28 +36,30 @@ class CheckCbasHealthStepTest extends BaseEmbeddedDbTest {
   @Test
   void doStepSuccess() {
     // setup
-    when(cbasService.checkHealth(any(), any()))
-        .thenReturn(new HealthCheckWorkspaceApps.Result(true, "cbas is healthy"));
+    when(wdsService.checkHealth(any(), any()))
+        .thenReturn(new HealthCheckWorkspaceApps.Result(true, "wds is healthy"));
+
+    StairwayTestUtils.constructCreateJobInputs(flightContext.getInputParameters());
 
     // do the step
-    CheckCbasHealthStep checkCbasHealthStep = new CheckCbasHealthStep(cbasService, samService);
-    StepResult result = checkCbasHealthStep.doStep(flightContext);
+    CheckWdsHealthStep checkWdsHealthStep = new CheckWdsHealthStep(wdsService, samService);
+    StepResult result = checkWdsHealthStep.doStep(flightContext);
 
     // make sure the step was a success
     assertEquals(StepStatus.STEP_RESULT_SUCCESS, result.getStepStatus());
   }
 
   @Test
-  void doStepUnhealthyCbas() {
+  void doStepUnhealthyWds() {
     // setup
-    when(cbasService.checkHealth(any(), any()))
+    when(wdsService.checkHealth(any(), any()))
         .thenReturn(new HealthCheckWorkspaceApps.Result(false, "wds is not healthy"));
 
     StairwayTestUtils.constructCreateJobInputs(flightContext.getInputParameters());
 
     // do the step
-    CheckCbasHealthStep checkCbasHealthStep = new CheckCbasHealthStep(cbasService, samService);
-    StepResult result = checkCbasHealthStep.doStep(flightContext);
+    CheckWdsHealthStep checkWdsHealthStep = new CheckWdsHealthStep(wdsService, samService);
+    StepResult result = checkWdsHealthStep.doStep(flightContext);
 
     // make sure the appropriate step status was returned
     assertEquals(StepStatus.STEP_RESULT_FAILURE_RETRY, result.getStepStatus());
@@ -64,8 +67,8 @@ class CheckCbasHealthStepTest extends BaseEmbeddedDbTest {
 
   @Test
   void undoStepSuccess() {
-    CheckCbasHealthStep checkCbasHealthStep = new CheckCbasHealthStep(cbasService, samService);
-    StepResult result = checkCbasHealthStep.undoStep(flightContext);
+    CheckWdsHealthStep checkWdsHealthStep = new CheckWdsHealthStep(wdsService, samService);
+    StepResult result = checkWdsHealthStep.undoStep(flightContext);
 
     assertEquals(StepStatus.STEP_RESULT_SUCCESS, result.getStepStatus());
   }
